refactor(http): move HTTP auth detection to pkg/brutus
Extract DetectHTTPAuthType from cmd/brutus/research.go into
pkg/brutus/http.go. Removes verbose logging parameter (library code
doesn't log). CLI wrapper routeHTTPWithAI now delegates to the library.
